+++ scripts/list_devices.py
@@ -1,3 +1,0 @@
-from approxeng.input.controllers import print_devices
-
-print_devices()